requirement.txt
async-timeout==4.0.3
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.34.4
botocore==1.34.162
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
reedsolo==1.7.0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.9.0
six==1.17.0
soupsieve==2.6
textual==2.1.2
tqdm==4.66.5
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.3.0
Werkzeug==3.1.3
zeroconf==0.136.0
